Se agrega la clase ProductosAComprar

+++ src/com/uno/carrito/Item.java
@@ -4,25 +4,30 @@ public class Item {
 
 	private int cantidad;
 	private Producto producto;
+	
 	public Item(Producto producto, int cantidad) {
 		super();
 		this.cantidad = cantidad;
 		this.producto = producto;
 	}
+	
 	public int getCantidad() {
 		return cantidad;
 	}
+	
 	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
+	
 	public Producto getProducto() {
 		return producto;
 	}
+	
 	public void setProducto(Producto producto) {
 		this.producto = producto;
 	}
 	
-	public float getTotal() {
+	public double getTotal() {
 		return cantidad * producto.getPrecio();
 	}
